update and fix error

+++ src/main/java/com/finance/alert_engine/dto/XauResponse.java
@@ -3,13 +3,11 @@ package com.finance.alert_engine.dto;
 import com.finance.alert_engine.util.DateUtil;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class XauResponse {
 
     private String name;
-    private BigDecimal price;
+    private Double price;
     private String symbol;
     private String timestamp = DateUtil.convertDate();
